Move more things to State.cpp/h, break out some stuff into a function
